15-1-llama-example/requirements.txt
sounddevice==0.4.7
numpy==1.26.4
soundfile==0.13.1
requests==2.31.0
openai-whisper==20240930
pyttsx3==2.98
